[adaptive sampling] Clean-up after previous refactoring
